代码/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.20
openai==1.57.0
python-dotenv==1.0.1
